backend/pyproject.toml
[project]
name = "eln-plus-plus"
version = "0.1.0"
description = "AI Co-Scientist for Biology — post-run multi-model ELN assistant"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.7",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "anthropic>=0.81.0",
    "langchain-core>=1.2.13",
    "langchain-anthropic>=1.3.3",
    "langgraph>=1.0.8",
    "langgraph-supervisor>=0.0.31",
    "qdrant-client>=1.9",
    "langchain-qdrant>=0.2",
    "pypdf>=6.7.1",
    "pdfplumber>=0.11.9",
    "python-pptx>=1.0.2",
    "python-docx>=1.1",
    "tiktoken>=0.12.0",
    "langchain-huggingface>=1.2.0",
    "sentence-transformers>=5.2.3",
    "langchain-text-splitters>=1.1.0",
    "httpx>=0.28.1",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "websockets>=12.0",
    "python-multipart>=0.0.9",
    "pandas>=2.3.3",
    "langchain-openai>=1.1.10",
    "langchain-google-genai>=4.2.0",
    "langchain-ollama>=0.2",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "opentelemetry-exporter-otlp>=1.20",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "aiosqlite>=0.22.1",
    "networkx>=3.0",
    "scipy>=1.11",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "statsmodels>=0.14",
    "pingouin>=0.5",
    "pillow>=10.0",
    "opencv-python>=4.8",
    "scikit-image>=0.22",
    "boto3>=1.42.76",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.10",
]
bench = [
    "datasets>=2.20",
    "huggingface_hub>=0.23",
]

[project.scripts]
api = "main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["eln"]

[tool.uv]
package = true

# Force CPU-only PyTorch — avoids downloading 2 GB of NVIDIA CUDA packages
# during Docker builds where a GPU is not present.
[tool.uv.sources]
torch = { index = "pytorch-cpu" }
torchvision = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[dependency-groups]
dev = [
    "ruff>=0.15.1",
]
